The Algo fills a tab in order to get the index of where to sort in stack A and the total number of ops it will require

+++ stacks/stacks_utils.c
@@ -6,7 +6,7 @@
 /*   By: sdummett <sdummett@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/17 20:57:34 by sdummett          #+#    #+#             */
-/*   Updated: 2021/09/03 12:25:37 by sdummett         ###   ########.fr       */
+/*   Updated: 2021/09/03 13:35:12 by sdummett         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,8 +22,8 @@ t_stacks	*create_stacks(int ac)
 	stacks->nb_elem_b = 0;
 	stacks->a = malloc(sizeof(int) * stacks->nb_elem_total);
 	stacks->b = malloc(sizeof(int) * stacks->nb_elem_total);
-	stacks->moves_a = malloc(sizeof(int) * stacks->nb_elem_total);
-	stacks->moves_b = malloc(sizeof(int) * stacks->nb_elem_total);
+	stacks->moves_a = malloc(sizeof(long int) * stacks->nb_elem_total);
+	stacks->moves_b = malloc(sizeof(long int) * stacks->nb_elem_total);
 	stacks->best_moves = malloc(sizeof(long int *) * 2);
 	stacks->best_moves[0] = malloc(sizeof(long int) * stacks->nb_elem_total);
 	stacks->best_moves[1] = malloc(sizeof(long int) * stacks->nb_elem_total);
